Implemented basic K nearest decks
Currently, distance is based on the relative frequency of a card
appearing across its class. Cards in a deck considered for the kth
nearest deck contribute nothing to the distance if they are in the list
of cards we have already seen. If the cards are commonly played they
contribute little to the distance, and if the cards are rarely played
by the class then they contribute a lot to the distance of the deck.
To improve on this, we should give higher weight to rarer cards that we
have seen the opponent play. e.g. if our opponent is a Warlock and
plays Golakka Crawler, they are probably playing a specialised version
of the deck, and so more general decks using more common cards
shouldn’t be considered as close as more specialised decks.